[hotfix] [rocksdb] Convert performance benchmarks to unit tests
